Abort where XARGS is required but is unset or empty
Presently, the "compress-file" unit defines the compress_file()
function, and the "phase-functions-sh" unit, the __dyn_clean() function.
Both functions require for the 'XARGS' environment variable to have been
set in order to behave correctly. Employ the ${param:?} form of
expansion so as to abort if the variable is either unset or empty.

Owing to a general lack of error checking, the principal benefit of this
change is that the diagnostics are improved for the case that the code
is not running in an appropriate operating environment.

BEFORE

ecompress-file: line 49: -0: command not found

AFTER

ecompress-file: line 49: XARGS: parameter null or not set

+++ bin/phase-functions.sh
@@ -324,7 +324,7 @@ __dyn_clean() {
 		printf '%s\0' "${PORTAGE_BUILDDIR}" \
 		| find0 -depth -type d -empty -print0 \
 		| while read -rd ''; do [[ ${REPLY} != "${WORKDIR}"?(/*) ]] && printf '%s\0' "${REPLY}"; done \
-		| ${XARGS} -0 rmdir --
+		| ${XARGS:?} -0 rmdir --
 	fi
 
 	# Do not bind this to doebuild defined DISTDIR; don't trust doebuild, and if mistakes are made it'll
